[CO-388] Fix x509 generation: throw when provided invalid config
e.g.
  - given an IP as AltDNS
  - given a non-positive expiry days

The former was a bit tricky. Actually, even if 'AltName' can be of
several shapes, they're only taken in account when they represent DNS
names. Therefore, I've adapted the generators to only generate valid
values and created a newtype 'Invalid a' that can be used to generate
invalid values of a and test that the function throws as expected.